added dropdown translator class
